+++ frontend/src/routes/AppRoutes.tsx
@@ -3,9 +3,9 @@ import { Chat } from '@/features/Chat/Chat';
 import { Pong } from '@/features/Pong/components/Pong';
 import { Index } from '@/features/Index/Index';
 import { DevAuth } from '@/features/DevAuth/DevAuth';
-import { useStoredCredential } from '@/hooks';
 import { useState } from 'react';
 import { UserPersonalData } from '@/features/DevAuth/AuthCard';
+import { useStoredCredential } from '@/features/DevAuth/hooks';
 
 export const AppRoutes = () => {
   // ブラウザが保持しているクレデンシャル
