+++ frontend/src/types.ts
@@ -1,0 +1,5 @@
+export type UserPersonalData = {
+  id: number;
+  email: string;
+  displayName: string;
+};
